+++ NEWS.md
@@ -4,6 +4,8 @@
   + *versionInfo* - gets version information of all components of LaBB-CAT
   + *saveParticipant* - add or update a participant record
   + *saveTranscript* - update transcript attributes
+- New LabbcatView function
+  + *readOnlyCategories* - lists attribute categories for read-only users
 
 Requires LaBB-CAT version 20230508.1434